+++ src/scripts/settings.ts
@@ -70,6 +70,7 @@ export const settingsStore = new Store({
     singleInstance: true,
     skipArtists: false,
     skippedArtists: [""],
+    startMinimized: false,
     staticWindowTitle: false,
     theme: "none",
     trayIcon: true,